requirements.txt
# Core dependencies
numpy>=1.19.2
opencv-python>=4.5.3
Pillow>=8.3.1
matplotlib>=3.4.3
scikit-image>=0.18.3
scikit-learn>=0.24.2
tqdm>=4.62.2
imagehash>=4.2.1

# Deep learning frameworks
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0
